....router....

+++ routes/board-route.js
@@ -12,7 +12,8 @@ const router = express.Router();
 const pugs = {
 	css: 'board', 
 	js: 'board', 
-	title: 'Express Board', 
+	title: 'Express Board',
+	tinyKey: process.env.TINY_KEY, 
 	headerTitle: 'Node/Express를 활용한 게시판' 
 }
 
@@ -43,9 +44,16 @@ router.get('/view/:id', async (req, res, next) => {
 			rs.filename = rs.orifile;
 			rs.src = imgExt.includes(extName(rs.savefile)) ? srcPath(rs.savefile) : null;
 		}
-		sql = 'SELECT id FROM board-ip WHERE bid=? AND ip=?';
+		console.log(ip.getClientIp(req));
+		sql = 'SELECT id FROM board_ip WHERE bid=? AND ip=?';
 		value = [rs.id, ip.getClientIp(req)];
 		r = await pool.query(sql, value);
+		sql = 'INSERT INTO board_ip SET bid=?, ip=?';
+		await pool.query(sql, value);
+		if(r[0].length == 0) {
+			sql = 'UPDATE board SET readnum = readnum + 1 WHERE id='+rs.id;
+			await pool.query(sql);
+		}
 		res.render('board/view', { ...pugs, rs });
 	}
 	catch(e) {
@@ -81,15 +89,15 @@ router.get(['/', '/list'], async (req, res, next) => {
 });
 
 router.get('/create', isUser, (req, res, next) => {
-	const pug = { ...pugs, tinyKey: process.env.TINY_KEY }
 	res.render('board/create', pugs);
 });
 
 router.post('/save', isUser, upload.single('upfile'), async (req, res, next) => {
 	try {
-		const { title, content, writer } = req.body;
-		let sql = 'INSERT INTO board SET title=?, content=?, writer=?, uid=?';
-		const value = [title, content, writer, req.session.user.id];
+		let sql, value, rs, r;
+		let { title, content, writer } = req.body;
+		sql = 'INSERT INTO board SET title=?, content=?, writer=?, uid=?';
+		value = [title, content, writer, req.session.user.id];
 		if(req.banExt) {
 			res.send(alert(`${req.banExt} 파일은 업로드 할 수 없습니다.`));
 		}
@@ -98,7 +106,7 @@ router.post('/save', isUser, upload.single('upfile'), async (req, res, next) => 
 				sql += ', orifile=?, savefile=?';
 				value.push(req.file.originalname, req.file.filename);
 			}
-			const r = await pool.query(sql, value);
+			r = await pool.query(sql, value);
 			res.redirect('/board');
 		}
 	}
@@ -113,11 +121,14 @@ router.get('/remove/:id', isUser, async (req, res, next) => {
 		sql = 'SELECT savefile FROM board WHERE id=? AND uid=?';
 		value = [req.params.id, req.session.user.id];
 		r = await pool.query(sql, value);
-		rs = r[0][0];
-		if(rs.savefile) await fs.remove(realPath(rs.savefile));
-		sql = 'DELETE FROM board WHERE id=? AND uid=?';
-		r = await pool.query(sql, value);
-		res.redirect('/board');
+		if(r[0].length == 0) res.send(alert('정상적인 접근이 아닙니다.'));
+		else {
+			rs = r[0][0];
+			if(rs.savefile) await fs.remove(realPath(rs.savefile));
+			sql = 'DELETE FROM board WHERE id=? AND uid=?';
+			r = await pool.query(sql, value);
+			res.redirect('/board');
+		}
 	}
 	catch(e) {
 		next(err(e.message));
@@ -126,15 +137,77 @@ router.get('/remove/:id', isUser, async (req, res, next) => {
 
 router.get('/change/:id', isUser, async (req, res, next) => {
 	try {
-
+		let sql, value, rs, r;
+		sql = 'SELECT * FROM board WHERE id=? AND uid=?';
+		value = [req.params.id, req.session.user.id];
+		r = await pool.query(sql, value);
+		if(r[0].length == 0) res.send(alert('정상적인 접근이 아닙니다.'));
+		else {
+			rs = r[0][0];
+			if(rs.savefile) {
+				rs.filename = rs.orifile;
+				rs.src = imgExt.includes(extName(rs.savefile)) ? srcPath(rs.savefile) : null;
+			}
+			res.render('board/change', { ...pugs, rs });
+		}
 	}
-	catch {
+	catch(e) {
 		next(err(e.message));
 	}
 });
 
-router.get('/api/remove/:id'. isUser. async (req, res, next) {
+router.get('/api/remove/:id', isUser, async (req, res, next) => {
+	try {
+		let sql, value, r, rs, id;
+		id = req.params.id;
+		sql = 'SELECT savefile FROM board WHERE id=? AND uid=?';
+		value = [req.params.id, req.session.user.id];
+		r = await pool.query(sql, value);
+		if(r[0].length == 0) res.json({ error: '삭제할 파일이 존재하지 않습니다.' })
+		else {
+			rs = r[0][0];
+			await fs.remove(realPath(rs.savefile));
+			sql = 'UPDATE board SET orifile=NULL, savefile=NULL WHERE id=? AND uid=?';
+			r = await pool.query(sql, value);
+			res.json({ code: 200 });
+		}
+	}
+	catch(e) {
+		next(err(e.message));
+	}
+});
 
+router.post('/update', isUser, upload.single('upfile'), async (req, res, next) => {
+	try {
+		let sql, value, rs, r;
+		let { title, content, writer, id } = req.body;
+		if(req.file) {
+			sql = 'SELECT savefile FROM board WHERE id=? AND uid=?';
+			value = [id, req.session.user.id];
+			r = await pool.query(sql, value);
+			if(r[0].length && r[0][0].savefile) {
+				await fs.remove(realPath(r[0][0].savefile));
+			}
+		}
+		sql = 'UPDATE board SET title=?, content=?, writer=? ';
+		value = [title, content, writer];
+		if(req.banExt) {
+			res.send(alert(`${req.banExt} 파일은 업로드 할 수 없습니다.`));
+		}
+		else {
+			if(req.file) {
+				sql += ', orifile=?, savefile=?';
+				value.push(req.file.originalname, req.file.filename);
+			}
+			sql += ' WHERE id=? AND uid=?';
+			value.push(id, req.session.user.id);
+			r = await pool.query(sql, value);
+			res.redirect('/board');
+		}
+	}
+	catch(e) {
+		next(err(e.message));
+	}
 });
 
 module.exports = router;
